Improvements to local calcuation
- Add `LocalSlot::addressed_from_sp`
- Use `u32` for local slot and local sizes calculation

+++ crates/winch/src/frame/mod.rs
@@ -6,12 +6,12 @@
 #[derive(Default)]
 pub(crate) struct Frame {
     /// The local area size
-    locals_size: u64,
+    locals_size: u32,
 }
 
 impl Frame {
     /// Allocate a new Frame
-    pub fn new(locals_size: u64) -> Self {
+    pub fn new(locals_size: u32) -> Self {
         Self { locals_size }
     }
 }
